feat: add initial ffi types, and preliminary work on the queue
## Type Issues
We used hsc2hs to generate basic FFI types. There wer some discrepencies
in c2hs, so we probably want to check out these types:
- All Word[N] types, c2hs wanted CULong and CUChar etc.
- CUIntPtr for the callback context, c2hs wanted CULong

## Threading the Queue into the C callback
We should verify that the foreign wrapper `makeCompletionCallback` works
I am not convinced it will magically turn a Haskell function into a C
Function Pointer.

## Tigerbeetle Semantics
We should verify how to access the data passed to the callback, and how
to parse the packets. It would also be nice to investigate some sort of 
stream interface for the Packet type.